Moved TopBar and Workflow so it can have its own CSS files. Imported a logo for the workflow and used it
