+++ client/src/components/setup/TopBar/TopBar.js
@@ -1,10 +1,11 @@
 import React, { Component } from 'react';
+import './TopBar.css';
 
 class TopBar extends Component {
     render() {
         return (
-            <div>
-                This is the TopBar component
+            <div className="p-30">
+                <img src={require("../../../assets/music-media-logo.png")} alt="Music Media Logo" />
             </div>
         );
     }
