+++ client/src/components/setup/Workflow/Workflow.css
@@ -1,0 +1,5 @@
+.form {
+    max-width: 850px;
+    margin: 0 auto;
+    background-color: white;
+}
